+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 dev:
   - ensure relay configuration inherits all configuration values as expected
+  - create strategies for builder bid
 
 1.7.6:
   - add User-Agent header to HTTP requests